# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * D-Tect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ExamType = (typeof ExamType)[keyof typeof ExamType];

export const ExamType = {
  mcq: "mcq",
  qa: "qa",
} as const;

export type ExamStatus = (typeof ExamStatus)[keyof typeof ExamStatus];

export const ExamStatus = {
  draft: "draft",
  active: "active",
  past: "past",
} as const;

export interface Exam {
  id: number;
  title: string;
  description?: string;
  type: ExamType;
  status: ExamStatus;
  secretCode?: string | null;
  createdAt: string;
  updatedAt: string;
  questionCount: number;
  submissionCount: number;
}

export type ExamWithQuestionsType =
  (typeof ExamWithQuestionsType)[keyof typeof ExamWithQuestionsType];

export const ExamWithQuestionsType = {
  mcq: "mcq",
  qa: "qa",
} as const;

export type ExamWithQuestionsStatus =
  (typeof ExamWithQuestionsStatus)[keyof typeof ExamWithQuestionsStatus];

export const ExamWithQuestionsStatus = {
  draft: "draft",
  active: "active",
  past: "past",
} as const;

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];

export const QuestionType = {
  mcq: "mcq",
  qa: "qa",
} as const;

export interface Question {
  id: number;
  examId: number;
  questionText: string;
  type: QuestionType;
  options?: string[] | null;
  correctAnswer?: string | null;
  orderIndex: number;
}

export interface ExamWithQuestions {
  id: number;
  title: string;
  description?: string;
  type: ExamWithQuestionsType;
  status: ExamWithQuestionsStatus;
  questions: Question[];
}

export interface ExamCode {
  code: string;
  examId: number;
}

export type CreateExamBodyType =
  (typeof CreateExamBodyType)[keyof typeof CreateExamBodyType];

export const CreateExamBodyType = {
  mcq: "mcq",
  qa: "qa",
} as const;

export interface CreateExamBody {
  title: string;
  description?: string;
  type: CreateExamBodyType;
}

export type UpdateExamBodyStatus =
  (typeof UpdateExamBodyStatus)[keyof typeof UpdateExamBodyStatus];

export const UpdateExamBodyStatus = {
  draft: "draft",
  active: "active",
  past: "past",
} as const;

export interface UpdateExamBody {
  title?: string;
  description?: string;
  status?: UpdateExamBodyStatus;
}

export interface JoinExamBody {
  code: string;
  studentId: number;
}

export type CreateQuestionBodyType =
  (typeof CreateQuestionBodyType)[keyof typeof CreateQuestionBodyType];

export const CreateQuestionBodyType = {
  mcq: "mcq",
  qa: "qa",
} as const;

export interface CreateQuestionBody {
  examId: number;
  questionText: string;
  type: CreateQuestionBodyType;
  options?: string[] | null;
  correctAnswer?: string | null;
  orderIndex: number;
}

export interface Answer {
  questionId: number;
  questionText: string;
  answer: string;
}

export interface Submission {
  id: number;
  examId: number;
  studentId: number;
  studentName: string;
  examTitle: string;
  answers: Answer[];
  score?: number | null;
  submittedAt: string;
}

export type CreateSubmissionBodyAnswersItem = {
  questionId: number;
  answer: string;
};

export interface CreateSubmissionBody {
  examId: number;
  studentId: number;
  answers: CreateSubmissionBodyAnswersItem[];
}

export interface UpdateScoreBody {
  /**
   * @minimum 0
   * @maximum 5
   */
  score: number;
}

export type ResourceType = (typeof ResourceType)[keyof typeof ResourceType];

export const ResourceType = {
  note: "note",
  link: "link",
  pdf: "pdf",
} as const;

export interface Resource {
  id: number;
  examId: number;
  title: string;
  content: string;
  type: ResourceType;
  url?: string | null;
  createdAt: string;
}

export type CreateResourceBodyType =
  (typeof CreateResourceBodyType)[keyof typeof CreateResourceBodyType];

export const CreateResourceBodyType = {
  note: "note",
  link: "link",
  pdf: "pdf",
} as const;

export interface CreateResourceBody {
  examId: number;
  title: string;
  content: string;
  type: CreateResourceBodyType;
  url?: string | null;
}

export interface Student {
  id: number;
  name: string;
  rollNumber: string;
  department: string;
  createdAt: string;
}

export interface CreateStudentBody {
  name: string;
  rollNumber: string;
  department: string;
}

export interface StatsOverview {
  totalExams: number;
  activeExams: number;
  totalStudents: number;
  totalSubmissions: number;
  averageScore: number;
  recentSubmissions: Submission[];
}

export interface StudentPerformance {
  studentId: number;
  studentName: string;
  rollNumber: string;
  department: string;
  totalExamsTaken: number;
  averageScore: number;
  submissions: Submission[];
}

export type ListExamsParams = {
  status?: ListExamsStatus;
};

export type ListExamsStatus =
  (typeof ListExamsStatus)[keyof typeof ListExamsStatus];

export const ListExamsStatus = {
  active: "active",
  past: "past",
  draft: "draft",
} as const;

export type ListQuestionsParams = {
  examId: number;
};

export type ListSubmissionsParams = {
  examId?: number;
};

export type ListResourcesParams = {
  examId?: number;
};